playback the awerb/ready/authority questions in the activity log
